[ch #153112486] Add assertions for Check attrs
The Check model's `last_ping` attribute is configured to accept None
values. The Check model's `n_pings` attribute is configured to set its
value to 0 incase no value is supplied. In this commit we create a new
test Check instance without specifying the `last_ping` and `n_pings`
attributes and check that the defaults are properly set.