[Feature #53112489] Render correct nag button state
The previous state of affairs was such that even after the nag mode of a
check is turned on, after a page re-render, the nag button state would
be displayed as off. This commit makes it such that if a check's nag
mode is on, then its state toggle button is rendered correctly.